qcacld-2.0: enable and tune BET
Enabling Beacon Early Termination in order
to speed up beacon processing drastically.
This will lead to power savings.

Also tune the BET wake interval to a value
of "8" to reduce the frequency of wakeups.
"8" is a safe choice since I could detect values
of higher than "10" to cause issues in some usecases.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>