Add 10 new Governors that cover all tastes and types of usage..
Read all about them here: https://forum.xda-developers.com/t/ref-guide-most-up-to-date-guide-on-cpu-governors-i-o-schedulers-and-more.3048957/

* Interactive Pro
A newer (modified) version of interactive which is optimized for devices such as the One Plus One.
It is a more efficient than the original Interactive because it continuously re-evaluates the load
of each CPU therefore allowing the CPU to scale efficiently.

* Impulse
An improved version of interactive modified by neobuddy89. Impulse aims to have a balance between
battery and performance just like interactive but has some tweaks to save battery.

* SmartassV2
Version 2 of the original smartass governor from Erasmux. The governor aim for an "ideal frequency",
and ramp up more aggressively towards this freq and less aggressive after.

* Alucard
Alucard is based on ondemand but has been heavily tweaked to bring better battery life and performance.
It has been known to be battery friendly without sacrificing much performance.

* PegasusQ
The Pegasus-q is a multi-core based on the Ondemand governor and governor with integrated hot-plugging.
It is quite stable and has the same battery life as ondemand.

* Dancedance
Based on conservative with some smartass features, it scales accordingly to conservatives laws. So it
will start from the bottom, take a load sample, if it's above the upthreshold, ramp up only one speed at
a time, and ramp down one at a time. The goal of this is to get the best battery life with decent performance.
It is a performance focused governor but also blends with some battery savings.

* Wheatley
Wheatley is a more performance orientated governor as it scales more aggressively than ondemand and sticks
with higher frequencies.

*Lionfish
The Lionfish governor combines traits of the conservative, ondemand, and interactive governors. It is designed
to maximize battery life without noticeably impacting performance. It responds quickly to heavy loads while
staying within the region of optimal CPU performance per watt. With moderate loads, it periodically votes to
raise, maintain, or decrease the frequency.

*Hyper
Hyper is an aggressive smart and smooth governor based on the Ondemand and is equipped with several features
of Ondemandx suspend profiles. It also has the fast_start deep_sleep variable and detection features. In addition,
the maximum frequency is in suspend mode 500Mhz or whatever the kernel developer sets it to. This is a more
smoothness oriented governor which means that it is good for performance, without sacrificing much battery life.

*Electron
Electron is a governor based on interactive from the latest MSM8994 CAF branch with more upstream improvements,
powersave bias, screen off max frequency, and some other tweaks to improve battery life without hindering performance.

Change-Id: I71f837b490812d10f06268cfde1f27aa2df70c76