Add 7 new hotplugs
A hotplug is a controller that chooses how many of your cpu cores are going to be used at once.
A governor governs at what frequency the core(s) are used, and a hotplug turnw off cores that do
not need to be utilized right now. This saves battery life, and also helps to keep the phone cool.

* Alucard
A great hotplugging driver by Alucard. It is known to be very battery friendly on devices.

* Autosmp
A highly-efficient hotplug driver by @mrg666, works in-sync with the CPU governor to enable off-line
cpu cores when the the CPU frequency reaches a high threshold and still more compute power is needed.

* Blu plug
Dynamic hotplug from eng.stk's shamu kernel with screenoff battery saving.

* Intelliplug
Great balance between battery life and performance. It is also a popular hotplug driver from faux123.

* Lazy plug
A battery orientated hotplug

* MSM hotplug
Great battery life, a custom qualcomm based hotplugging driver by myflux. It is a popular choice for many users.

* Thunderplug
A mature load-based hotplug driver with many tunables written from ground up by varun.chitre15. This hotplug is
optimized for octa-core devices and also has support for 64bit CPUs.